requirements.txt
flask==2.1.2
python-dotenv
boto3
pyjwt[cypto]
